+++ src/models/Hand.js
@@ -1,15 +1,14 @@
-export default class Hand {
+import TileCollection from "./TileCollection.js";
+
+export default class Hand extends TileCollection {
   constructor() {
-    this.tiles = [];
+    super();
   }
 
   get playableTiles() {
     return this.sortTiles(this.tiles.filter(tile => tile.type === "playable"));
   }
 
-  getBySuit(suit) {
-    return this.sortTiles(this.tiles.filter(tile => tile.suit === suit));
-  }
 
   getLongestInSuit() {
     const suits = ["C", "●", "┇"];
@@ -41,39 +40,9 @@ export default class Hand {
     lines.push(`\n      Points: ${this.pointsTiles.map(t => t.toString()).join("")}`);
     return lines.join("");
   }
-
-  sortTiles(tiles) {
-    const suitOrder = { C: 0, "●": 1, "┇": 2 };
-
-    return [...tiles].sort((a, b) => {
-      if (suitOrder[a.suit] !== suitOrder[b.suit]) {
-        return suitOrder[a.suit] - suitOrder[b.suit];
-      }
-      return a.value - b.value;
-    });
-  }
   
   get pointsTiles() {
     return this.tiles.filter(tile => tile.type === "points");
-  }
-
-  addTile(tile) {
-    this.tiles.push(tile);
-  }
-
-  removeTile(tile) {
-    if (!Array.isArray(this.tiles)) {
-      console.error('tiles is not an array');
-      return false;
-    }
-
-    const idx = this.tiles.indexOf(tile);
-    if (idx >= 0) {
-      this.tiles.splice(idx, 1);
-      return true;
-    }
-
-    return false;
   }
 
   toString() {
